+++ designer/client/src/components/modals/Testing/TestingDialog.tsx
@@ -1,17 +1,14 @@
 import { Box } from "@mui/material";
 import type { WindowContentProps } from "@touk/window-manager";
-import type { ElementType, ReactElement} from "react";
-import React, { useEffect, useState } from "react";
-import { useSelector } from "react-redux";
+import type { ElementType, ReactElement } from "react";
+import React from "react";
 
-import { getTestCapabilities } from "../../../reducers/selectors/graph";
 import type { WindowKind } from "../../../windowManager";
 import { WindowContent } from "../../../windowManager";
 import { ContentSize } from "../../graph/node-modal/node/ContentSize";
 import { WindowHeaderIconStyled } from "../../graph/node-modal/nodeDetails/NodeDetailsStyled";
 import { NodeDocs } from "../../graph/node-modal/nodeDetails/SubHeader";
-import type { FormValue, TouchedValue } from "./TestingForm";
-import { TestingForm, TestType } from "./TestingForm";
+import { TestingForm } from "./TestingForm";
 
 type DocsLink = {
     url: string;
@@ -19,8 +16,6 @@ type DocsLink = {
 };
 
 export type TestingViewParams = {
-    confirmText?: string;
-    cancelText?: string;
     Icon?: ElementType;
     docs?: DocsLink;
     // may contain a ::form-fields or ::form-field{name=""} directives
@@ -28,57 +23,25 @@ export type TestingViewParams = {
 };
 
 export interface TestingData {
-    view: TestingViewParams;
+    viewParams: TestingViewParams;
 }
 
 function TestingDialog(props: WindowContentProps<WindowKind, TestingData>): ReactElement {
     const { data, close } = props;
     const {
-        meta: { view },
+        meta: { viewParams },
         kind,
     } = data;
-
-    const testCapabilities = useSelector(getTestCapabilities);
-
-    const availabilityMap: Record<TestType, boolean> = {
-        [TestType.withParameters]: testCapabilities.canTestWithForm,
-        [TestType.withGeneratedData]: testCapabilities.canGenerateTestData && testCapabilities.canBeTested,
-    };
-    const availableTestTypes = Object.entries(availabilityMap)
-        .filter(([_, isAvailable]) => isAvailable)
-        .map(([key]) => key as TestType);
-
-    const [testType, setState] = useValidatedLocalStorage<FormValue>(
-        "selectedTestTypeState",
-        { testType: availableTestTypes[0] },
-        (stored: FormValue) => availableTestTypes.includes(TestType[stored.testType]),
-    );
-    const [touched, setTouched] = useState<TouchedValue>({
-        testType: false,
-    });
-    const onChange = (value: FormValue) => {
-        setState(value);
-    };
-    const handleSetTouched = (touched: TouchedValue) => {
-        setTouched(touched);
-    };
 
     return (
         <WindowContent
             {...props}
-            icon={<WindowHeaderIconStyled as={view.Icon} type={kind} />}
-            subheader={<NodeDocs name={view.docs?.label} href={view.docs?.url} />}
+            icon={<WindowHeaderIconStyled as={viewParams.Icon} type={kind} />}
+            subheader={<NodeDocs name={viewParams.docs?.label} href={viewParams.docs?.url} />}
         >
             <ContentSize>
                 <Box mx={3}>
-                    <TestingForm
-                        value={testType}
-                        onChange={onChange}
-                        touched={touched}
-                        handleSetTouched={handleSetTouched}
-                        testingData={props.data}
-                        closeDialog={close}
-                    />
+                    <TestingForm testingData={props.data} closeDialog={close} />
                 </Box>
             </ContentSize>
         </WindowContent>
@@ -86,24 +49,3 @@ function TestingDialog(props: WindowContentProps<WindowKind, TestingData>): Reac
 }
 
 export default TestingDialog;
-
-function useValidatedLocalStorage<T>(key: string, initialValue: T, validate: (value: any) => boolean) {
-    const [state, setState] = useState<T>(() => {
-        try {
-            const storedValue = localStorage.getItem(key);
-            if (storedValue !== null) {
-                const parsedValue = JSON.parse(storedValue);
-                return validate(parsedValue) ? parsedValue : initialValue;
-            }
-        } catch (error) {
-            console.error("Error reading localStorage:", error);
-        }
-        return initialValue;
-    });
-
-    useEffect(() => {
-        localStorage.setItem(key, JSON.stringify(state));
-    }, [key, state]);
-
-    return [state, setState] as const;
-}
